product/juno: Fix required MISRA defects rule 10.3
MSIRA 10.3 defects relating to product files will be addressed
in this patch.

+++ product/juno/module/juno_cdcel937/src/juno_cdcel937.h
@@ -60,7 +60,7 @@ struct field {
  */
 #define FIELD(_name, _mask) \
     static const struct field _name = { \
-        .reg_idx = _name##_IDX, \
+        .reg_idx = (uint32_t)_name##_IDX, \
         .mask = _mask, \
         .pos = __builtin_ctz(_mask), \
     }
